Move MUC invite modal to its own file
